apps/project/api_endpoints/service/ServiceList/serializers.py
from rest_framework import serializers
from apps.project.models import Service


class ServiceSerializer(serializers.ModelSerializer):
    class Meta:
        model = Service
        fields = ['id', 'name']
