requirements/base.txt
Django==4.2
environs[django]==9.5.0
django-jazzmin==2.6.0
drf-yasg==1.21.5
django-phonenumber-field==7.0.2
phonenumbers==8.13.9
Pillow==9.5.0
djangorestframework==3.14.0
django-filter==23.1